requirements.txt
numpy
pillow

torch
torchvision

transformers

opencv-python
mediapipe

